+++ server/src/models/index.js
@@ -1,11 +1,10 @@
 import fs from 'fs';
 import path from 'path';
 import Sequelize from 'sequelize';
-import configAll from '../../config/config.json';
+import configAll from '../../config/database_config.json';
 
 const basename = path.basename(module.filename);
 const env = process.env.NODE_ENV || 'development';
-// const env = process.env.NODE_ENV;
 const config = configAll[env];
 const db = {};
 
@@ -17,10 +16,6 @@ if (config.use_env_variable) {
     config.database, config.username, config.password, config
   );
 }
-
-// const sequelize = new Sequelize(
-//   config.database, config.username, config.password, config
-// );
 
 fs
   .readdirSync(__dirname)
